🧪 Reduce excessive test coverage and eliminate redundant tests
- Reduced total test functions from 1,181 to 985 (16.6% reduction)
- Consolidated tutorial tests: merged test_docker_tutorial.py and test_tutorial_executor.py into test_tutorial.py
- Merged performance tests: consolidated test_performance_benchmark.py into test_performance.py
- Unified services tests: combined all test_services_*.py files into single test_services.py
- Reduced utility test files by ~50%: test_utils.py, test_cache.py, test_client.py
- Reduced admin/articles test files by ~40%: test_admin.py, test_articles.py
- Eliminated redundant parameter combinations and excessive edge case testing
- Maintained 90%+ coverage for critical code paths while improving test maintainability
- Focused on high-value tests covering core functionality and essential error conditions

Fixes #422
