+++ tests/test_performance.py
@@ -2,7 +2,7 @@
 
 import asyncio
 import time
-from unittest.mock import Mock, patch
+from unittest.mock import AsyncMock, Mock, patch
 
 import pytest
 
@@ -14,6 +14,7 @@ from youtrack_cli.performance import (
     get_performance_monitor,
     performance_timer,
 )
+from youtrack_cli.performance_benchmark import FieldSelectionBenchmark, run_benchmark
 
 
 @pytest.mark.unit
@@ -479,3 +480,127 @@ class TestPerformanceIntegration:
         create_summary = monitor.summary(operation="create_issue")
         assert create_summary["total_operations"] == 1  # Only successful
         assert create_summary["failed_operations"] == 1
+
+
+# === CONSOLIDATED BENCHMARK TESTS FROM test_performance_benchmark.py ===
+
+
+class TestFieldSelectionBenchmark:
+    """Test FieldSelectionBenchmark class."""
+
+    def test_init(self):
+        """Test benchmark initialization."""
+        mock_auth_manager = Mock()
+
+        with patch("youtrack_cli.performance_benchmark.IssueManager") as mock_issue_manager:
+            with patch("youtrack_cli.performance_benchmark.get_field_selector") as mock_field_selector:
+                benchmark = FieldSelectionBenchmark(mock_auth_manager)
+
+                assert benchmark.auth_manager == mock_auth_manager
+                mock_issue_manager.assert_called_once_with(mock_auth_manager)
+                mock_field_selector.assert_called_once()
+
+    @pytest.mark.asyncio
+    async def test_benchmark_profile_performance_success(self):
+        """Test successful benchmark profile performance."""
+        mock_auth_manager = Mock()
+
+        with patch("youtrack_cli.performance_benchmark.IssueManager") as mock_issue_manager_cls:
+            with patch("youtrack_cli.performance_benchmark.get_field_selector"):
+                mock_issue_manager = Mock()
+                mock_issue_manager.list_issues = AsyncMock(return_value={"status": "success", "issues": []})
+                mock_issue_manager_cls.return_value = mock_issue_manager
+
+                benchmark = FieldSelectionBenchmark(mock_auth_manager)
+
+                with patch("youtrack_cli.performance_benchmark.logger") as mock_logger:
+                    time_values = [i * 0.1 for i in range(50)]
+                    with patch("time.time", side_effect=time_values):
+                        results = await benchmark.benchmark_profile_performance(
+                            project_id="TEST", query="test query", sample_size=10
+                        )
+
+                assert "minimal" in results
+                assert "standard" in results
+                assert "full" in results
+
+                for profile in ["minimal", "standard", "full"]:
+                    assert "avg_time" in results[profile]
+                    assert "min_time" in results[profile]
+                    assert "max_time" in results[profile]
+                    assert "times" in results[profile]
+                    assert len(results[profile]["times"]) == 5
+
+                mock_logger.info.assert_called()
+
+    @pytest.mark.asyncio
+    async def test_run_single_benchmark_success(self):
+        """Test successful single benchmark run."""
+        mock_auth_manager = Mock()
+
+        with patch("youtrack_cli.performance_benchmark.IssueManager") as mock_issue_manager_cls:
+            with patch("youtrack_cli.performance_benchmark.get_field_selector"):
+                mock_issue_manager = Mock()
+                mock_issue_manager.list_issues = AsyncMock(return_value={"status": "success", "issues": []})
+                mock_issue_manager_cls.return_value = mock_issue_manager
+
+                benchmark = FieldSelectionBenchmark(mock_auth_manager)
+
+                with patch("time.time", side_effect=[0.0, 0.5]):
+                    duration = await benchmark._run_single_benchmark("TEST", "query", "minimal", 50)
+
+                assert duration == 0.5
+                mock_issue_manager.list_issues.assert_called_once_with(
+                    project_id="TEST", query="query", field_profile="minimal", page_size=50, top=50
+                )
+
+    def test_analyze_performance_gains_success(self):
+        """Test performance gains analysis with all profiles."""
+        mock_auth_manager = Mock()
+
+        with patch("youtrack_cli.performance_benchmark.IssueManager"):
+            with patch("youtrack_cli.performance_benchmark.get_field_selector"):
+                benchmark = FieldSelectionBenchmark(mock_auth_manager)
+
+                results = {
+                    "minimal": {"avg_time": 0.5},
+                    "standard": {"avg_time": 0.8},
+                    "full": {"avg_time": 1.0},
+                }
+
+                with patch("youtrack_cli.performance_benchmark.logger"):
+                    improvements = benchmark.analyze_performance_gains(results)
+
+                assert "minimal_vs_full" in improvements
+                assert "standard_vs_full" in improvements
+                assert "estimated_data_reduction" in improvements
+
+                assert abs(improvements["minimal_vs_full"] - 50.0) < 0.01
+                assert abs(improvements["standard_vs_full"] - 20.0) < 0.01
+                assert improvements["estimated_data_reduction"] == 75.0
+
+
+class TestRunBenchmark:
+    """Test run_benchmark function."""
+
+    @pytest.mark.asyncio
+    async def test_run_benchmark_success(self):
+        """Test successful benchmark run."""
+        mock_auth_manager = Mock()
+
+        with patch("youtrack_cli.performance_benchmark.FieldSelectionBenchmark") as mock_benchmark_cls:
+            mock_benchmark = Mock()
+            mock_benchmark.benchmark_profile_performance = AsyncMock(
+                return_value={
+                    "minimal": {"avg_time": 0.5, "min_time": 0.4, "max_time": 0.6},
+                    "full": {"avg_time": 1.0, "min_time": 0.9, "max_time": 1.1},
+                }
+            )
+            mock_benchmark.print_benchmark_report = Mock()
+            mock_benchmark_cls.return_value = mock_benchmark
+
+            await run_benchmark(mock_auth_manager, project_id="TEST", sample_size=25)
+
+            mock_benchmark_cls.assert_called_once_with(mock_auth_manager)
+            mock_benchmark.benchmark_profile_performance.assert_called_once_with(project_id="TEST", sample_size=25)
+            mock_benchmark.print_benchmark_report.assert_called_once()
